+++ src/api/client.js
@@ -18,14 +18,15 @@ export default client
  * Creates an entity associated with the current output
  * @param {string} endpoint the endpoint that handles the entity
  * @param {object} payload the payload for the object
+ * @param {object} cfg The axios config object
  * @returns {Array} A list containing the response and error respectivly
  */
-export async function createEntity(endpoint, payload) {
+export async function createEntity(endpoint, payload, cfg = null) {
   let res = null
   let error = null
 
   try {
-    res = await client.post(endpoint, payload)
+    res = await client.post(endpoint, payload, cfg)
   } catch (e) {
     error = e
   }
@@ -47,10 +48,10 @@ export async function createEntity(endpoint, payload) {
  * Update an entity based on a given endpoint
  * @param {String} endpoint 
  * @param {PayloadWithId} payload The payload to update
- * @param {Array}
- * @returns 
+ * @param {object} cfg The axios config object
+ * @returns {Array}
  */
-export async function updateEntity(endpoint, payload) {
+export async function updateEntity(endpoint, payload, cfg = null) {
   if (!payload.hasOwnProperty('id')) {
     throw new Error("Cannot update entity for endpoint 'endpoint' an id MUST be supplied")
   }
@@ -67,7 +68,7 @@ export async function updateEntity(endpoint, payload) {
   if (!error) {
     // Dispatch and mutate exsisting state with what we posted
     mutate(endpoint, (current) => {
-      return current.map(item => item.id == payload.id ? payload : item)
+      return current.map(item => item.id === payload.id ? payload : item)
     })
 
     mutate(`${endpoint}/${payload.id}`, payload)
