+++ src/components/entry/SystemMapEntry/EyeSpaceProfileEntry.stories.js
@@ -8,9 +8,11 @@ export default {
 
 const Template = (args) => <SystemMapEntry {...args}/>;
 
-export const Primary = Template.bind({});
+export const Primary = Template.bind({})
 
 Primary.args = {
-  text: "Wonda",
+  text: "Melissa",
+  subText: "Female",
+  age: "Age LOL",
   img: 'https://i.pinimg.com/736x/6c/12/a8/6c12a8359754acb9f176c9155637c912.jpg',
 }
