+++ src/components/entry/SystemMapEntry/SystemMapEntry.jsx
@@ -1,21 +1,30 @@
 import React from 'react';
 import Container from "../../generic/container/Container";
-import { Typography } from '@material-ui/core';
+import GenericButton from "../../generic/buttons/GenericButton"
+import { Typography, Grid } from '@material-ui/core';
 import PropTypes from 'prop-types'; 
 
 function SystemMapEntry(props){
-    return (<Container>
-        {props.img}
-        <Typography variant='h6'>{props.text}</Typography>
-    </Container>)
+    return (<GenericButton style={{ marginLeft: '1rem', marginTop: '1rem' }} >
+    <Grid container>
+      <Grid item xs='6'>
+        <Typography variant='caption'>{props.text}</Typography>
+        <Typography variant='subtitle1'>{props.subText}<br />{props.age}</Typography>
+      </Grid>
+      <Grid item xs='6' style={{ paddingLeft: '1rem' }}>
+        <Container>
+          <img src={props.img} style={{maxHeight: '120px'}} alt="profile picture"/>
+        </Container>
+      </Grid>
+    </Grid>
+  </GenericButton>)
 }
 
 SystemMapEntry.propTypes = {
-    img: PropTypes.string,
-    text: PropTypes.string
-}
+  text: PropTypes.string,
+  subText: PropTypes.string,
+  age: PropTypes.string,
+  img: PropTypes.string 
+} 
 
 export default SystemMapEntry
-        
-
-
